Improved single iFrame selection
Added functionality for visual selection of game. iFrame repositioned to
top of site. Currently working on jquery "back to top" with animation to
bring user back to the iframe at the top of the page when they select a
game. Added first screenshot of game "CY" to text effect and make
adjustments accordingly before adding other screenshots